New Camera and Mana Ball fix
I added cinemachine to the project making new camera system that is cool. Struggling to include camera boundaries tho. And added on loading of a scene it will plop a player down in that scene useful for build and testing. Also fixed the mana ui ball not updating whne mana spent.